# docs/api/openapi.json
{
  "components" : {
    "schemas" : {
...
      "ChangerMotDePasseRequestDto" : {
        "properties" : {
          "ancienMotDePasse" : {
            "minLength" : 1,
            "type" : "string"
          },
          "nouveauMotDePasse" : {
            "maxLength" : 255,
            "minLength" : 8,
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "CompteCreeDto" : {
        "properties" : {
          "compte" : {
            "$ref" : "#/components/schemas/UtilisateurCompteDto"
          },
          "motDePasseTemporaire" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "CreerUtilisateurRequestDto" : {
        "properties" : {
          "email" : {
            "maxLength" : 320,
            "minLength" : 0,
            "type" : "string"
          },
          "nomComplet" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string"
          },
          "roles" : {
            "items" : {
              "enum" : [
                "SUPER_ADMIN",
                "ADMIN",
                "DIRECTION",
                "GESTIONNAIRE",
                "ENSEIGNANT",
                "PARENT",
                "ELEVE"
              ],
              "type" : "string"
            },
            "minItems" : 1,
            "type" : "array",
            "uniqueItems" : true
          },
          "siteId" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ModifierRolesRequestDto" : {
        "properties" : {
          "roles" : {
            "items" : {
              "enum" : [
                "SUPER_ADMIN",
                "ADMIN",
                "DIRECTION",
                "GESTIONNAIRE",
                "ENSEIGNANT",
                "PARENT",
                "ELEVE"
              ],
              "type" : "string"
            },
            "minItems" : 1,
            "type" : "array",
            "uniqueItems" : true
          }
        },
        "type" : "object"
      },
...
      "MotDePasseTemporaireDto" : {
        "properties" : {
          "motDePasseTemporaire" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "PageReponseUtilisateurCompteDto" : {
        "properties" : {
          "contenu" : {
            "items" : {
              "$ref" : "#/components/schemas/UtilisateurCompteDto"
            },
            "type" : "array"
          },
          "page" : {
            "format" : "int32",
            "type" : "integer"
          },
          "taille" : {
            "format" : "int32",
            "type" : "integer"
          },
          "totalElements" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "PageReponseUtilisateurResumeDto" : {
        "properties" : {
          "contenu" : {
            "items" : {
              "$ref" : "#/components/schemas/UtilisateurResumeDto"
            },
            "type" : "array"
          },
          "page" : {
            "format" : "int32",
            "type" : "integer"
          },
          "taille" : {
            "format" : "int32",
            "type" : "integer"
          },
          "totalElements" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "UtilisateurCompteDto" : {
        "properties" : {
          "actif" : {
            "type" : "boolean"
          },
          "dateCreation" : {
            "format" : "date-time",
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "motDePasseAChanger" : {
            "type" : "boolean"
          },
          "nomComplet" : {
            "type" : "string"
          },
          "roles" : {
            "items" : {
              "enum" : [
                "SUPER_ADMIN",
                "ADMIN",
                "DIRECTION",
                "GESTIONNAIRE",
                "ENSEIGNANT",
                "PARENT",
                "ELEVE"
              ],
              "type" : "string"
            },
            "type" : "array",
            "uniqueItems" : true
          },
          "siteId" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "UtilisateurResumeDto" : {
        "properties" : {
          "actif" : {
            "type" : "boolean"
          },
          "email" : {
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "motDePasseAChanger" : {
            "type" : "boolean"
          },
          "nomComplet" : {
            "type" : "string"
          },
          "superAdmin" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      }
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "in" : "header",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "API de gestion scolaire Edukeys, éditée par Nova Digital (Togo).",
    "title" : "Edukeys — API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/utilisateurs" : {
      "get" : {
        "operationId" : "listerUtilisateursPlateforme",
        "parameters" : [
          {
            "in" : "query",
            "name" : "page",
            "required" : false,
            "schema" : {
              "format" : "int32",
              "type" : "integer"
            }
          },
          {
            "in" : "query",
            "name" : "size",
            "required" : false,
            "schema" : {
              "format" : "int32",
              "type" : "integer"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageReponseUtilisateurResumeDto"
                }
              }
            },
            "description" : "Page de comptes"
          },
          "403" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageReponseUtilisateurResumeDto"
                }
              }
            },
            "description" : "Permission UTILISATEUR_GERER_PLATEFORME requise"
          }
        },
        "summary" : "Liste paginée des comptes utilisateurs de tous les établissements (administration de plateforme)",
        "tags" : [
          "Utilisateurs"
        ]
      },
      "post" : {
        "operationId" : "creerUtilisateur",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreerUtilisateurRequestDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CompteCreeDto"
                }
              }
            },
            "description" : "Compte créé, avec le mot de passe temporaire de son titulaire (retourné une seule fois)"
          },
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CompteCreeDto"
                }
              }
            },
            "description" : "Cet email est déjà utilisé sur la plateforme (message unique, y compris pour un compte de plateforme)"
          },
          "422" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CompteCreeDto"
                }
              }
            },
            "description" : "Aucun rôle fourni, ou SUPER_ADMIN demandé (rôle de plateforme, non attribuable ici)"
          }
        },
        "summary" : "Crée un compte dans l'établissement courant",
        "tags" : [
          "Utilisateurs"
        ]
      }
    },
...
    "/api/v1/utilisateurs/moi/mot-de-passe" : {
      "post" : {
        "operationId" : "changerMonMotDePasse",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ChangerMotDePasseRequestDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "Mot de passe changé"
          },
          "401" : {
            "description" : "Ancien mot de passe incorrect"
          }
        },
        "summary" : "Change le mot de passe du compte authentifié",
        "tags" : [
          "Utilisateurs"
        ]
      }
    },
    "/api/v1/utilisateurs/mon-etablissement" : {
      "get" : {
        "operationId" : "listerUtilisateursMonEtablissement",
        "parameters" : [
          {
            "in" : "query",
            "name" : "page",
            "required" : false,
            "schema" : {
              "format" : "int32",
              "type" : "integer"
            }
          },
          {
            "in" : "query",
            "name" : "size",
            "required" : false,
            "schema" : {
              "format" : "int32",
              "type" : "integer"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageReponseUtilisateurCompteDto"
                }
              }
            },
            "description" : "Page de comptes de l'établissement courant"
          }
        },
        "summary" : "Liste paginée des comptes de l'établissement courant",
        "tags" : [
          "Utilisateurs"
        ]
      }
    },
    "/api/v1/utilisateurs/mon-etablissement/{id}" : {
      "get" : {
        "operationId" : "obtenirUtilisateurMonEtablissement",
        "parameters" : [
          {
            "in" : "path",
            "name" : "id",
            "required" : true,
            "schema" : {
              "format" : "uuid",
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UtilisateurCompteDto"
                }
              }
            },
            "description" : "Compte de l'établissement courant"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UtilisateurCompteDto"
                }
              }
            },
            "description" : "Aucun compte actif portant cet identifiant dans l'établissement courant"
          }
        },
        "summary" : "Compte de l'établissement courant par identifiant",
        "tags" : [
          "Utilisateurs"
        ]
      }
    },
    "/api/v1/utilisateurs/{id}" : {
      "delete" : {
        "operationId" : "desactiverUtilisateur",
        "parameters" : [
          {
            "in" : "path",
            "name" : "id",
            "required" : true,
            "schema" : {
              "format" : "uuid",
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "204" : {
            "description" : "Affectation désactivée"
          },
          "404" : {
            "description" : "Aucun compte actif portant cet identifiant dans l'établissement courant"
          },
          "422" : {
            "description" : "Tentative de désactiver son propre compte, ou dernier ADMIN actif de l'établissement"
          }
        },
        "summary" : "Désactive (logiquement) l'affectation d'un compte à l'établissement courant",
        "tags" : [
          "Utilisateurs"
        ]
      }
    },
    "/api/v1/utilisateurs/{id}/mot-de-passe-temporaire" : {
      "post" : {
        "operationId" : "regenererMotDePasseTemporaire",
        "parameters" : [
          {
            "in" : "path",
            "name" : "id",
            "required" : true,
            "schema" : {
              "format" : "uuid",
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MotDePasseTemporaireDto"
                }
              }
            },
            "description" : "Nouveau mot de passe temporaire (retourné une seule fois)"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MotDePasseTemporaireDto"
                }
              }
            },
            "description" : "Aucun compte actif portant cet identifiant dans l'établissement courant"
          }
        },
        "summary" : "Régénère un mot de passe temporaire pour un compte de l'établissement courant",
        "tags" : [
          "Utilisateurs"
        ]
      }
    },
    "/api/v1/utilisateurs/{id}/reactiver" : {
      "post" : {
        "operationId" : "reactiverUtilisateur",
        "parameters" : [
          {
            "in" : "path",
            "name" : "id",
            "required" : true,
            "schema" : {
              "format" : "uuid",
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "204" : {
            "description" : "Affectation réactivée"
          },
          "404" : {
            "description" : "Aucun compte portant cet identifiant dans l'établissement courant"
          }
        },
        "summary" : "Réactive l'affectation d'un compte à l'établissement courant",
        "tags" : [
          "Utilisateurs"
        ]
      }
    },
    "/api/v1/utilisateurs/{id}/roles" : {
      "put" : {
        "operationId" : "remplacerRolesUtilisateur",
        "parameters" : [
          {
            "in" : "path",
            "name" : "id",
            "required" : true,
            "schema" : {
              "format" : "uuid",
              "type" : "string"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ModifierRolesRequestDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "Rôles remplacés"
          },
          "404" : {
            "description" : "Aucun compte actif portant cet identifiant dans l'établissement courant"
          },
          "422" : {
            "description" : "Aucun rôle fourni, SUPER_ADMIN demandé, ou tentative de modifier ses propres rôles"
          }
        },
        "summary" : "Remplace les rôles d'un compte de l'établissement courant",
        "tags" : [
          "Utilisateurs"
        ]
      }
    }
  },
  "security" : [
    {
      "bearerAuth" : [ ]
    }
  ],
  "servers" : [
    {
      "description" : "Generated server url",
      "url" : "http://localhost"
    }
  ]
}